Move check_for_origin_id() to MotionCreateForwarded.
