+++ openslides_backend/action/actions/motion/create_base.py
@@ -51,33 +51,6 @@ class MotionCreateBase(
                 "No matching default workflow defined on this meeting"
             )
 
-    def check_for_origin_id(self, instance: Dict[str, Any]) -> None:
-        if instance.get("origin_id"):
-            meeting = self.datastore.get(
-                FullQualifiedId(Collection("meeting"), instance["meeting_id"]),
-                ["committee_id"],
-            )
-            forwarded_from = self.datastore.get(
-                FullQualifiedId(Collection("motion"), instance["origin_id"]),
-                ["meeting_id"],
-            )
-            forwarded_from_meeting = self.datastore.get(
-                FullQualifiedId(Collection("meeting"), forwarded_from["meeting_id"]),
-                ["committee_id"],
-            )
-            committee = self.datastore.get(
-                FullQualifiedId(
-                    Collection("committee"), forwarded_from_meeting["committee_id"]
-                ),
-                ["forward_to_committee_ids"],
-            )
-            if meeting["committee_id"] not in committee.get(
-                "forward_to_committee_ids", []
-            ):
-                raise ActionException(
-                    f"Committee id {meeting['committee_id']} not in {committee.get('forward_to_committee_ids', [])}"
-                )
-
     def create_submitters(self, instance: Dict[str, Any]) -> None:
         submitter_ids = instance.pop("submitter_ids", None)
         if not submitter_ids:
